moved int_admin_ometa to a dedicated module
