+++ ingestion/tests/integration/conftest.py
@@ -4,7 +4,7 @@ import sys
 
 import pytest
 
-from .integration_base import int_admin_ometa
+from _openmetadata_testutils.ometa import int_admin_ometa
 
 if not sys.version_info >= (3, 9):
     collect_ignore = ["trino"]
