+++ ingestion/tests/integration/ometa/test_ometa_table_api.py
@@ -22,6 +22,7 @@ from unittest.mock import patch
 import pytest
 from pydantic import ValidationError
 
+from _openmetadata_testutils.ometa import int_admin_ometa
 from metadata.generated.schema.api.data.createDatabase import CreateDatabaseRequest
 from metadata.generated.schema.api.data.createDatabaseSchema import (
     CreateDatabaseSchemaRequest,
@@ -73,8 +74,6 @@ from metadata.generated.schema.type.basic import (
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.generated.schema.type.usageRequest import UsageRequest
 from metadata.ingestion.ometa.client import REST
-
-from ..integration_base import int_admin_ometa
 
 BAD_RESPONSE = {
     "data": [
