+++ ingestion/tests/integration/ometa/test_ometa_user_api.py
@@ -15,12 +15,11 @@ import logging
 import time
 from unittest import TestCase
 
+from _openmetadata_testutils.ometa import int_admin_ometa
 from metadata.generated.schema.api.teams.createTeam import CreateTeamRequest
 from metadata.generated.schema.api.teams.createUser import CreateUserRequest
 from metadata.generated.schema.entity.teams.team import Team, TeamType
 from metadata.generated.schema.entity.teams.user import User
-
-from ..integration_base import int_admin_ometa
 
 
 class OMetaUserTest(TestCase):
