+++ ingestion/tests/integration/s3/conftest.py
@@ -17,12 +17,12 @@ import pytest
 import yaml
 from minio import Minio
 
+from _openmetadata_testutils.ometa import OM_JWT, int_admin_ometa
 from metadata.generated.schema.entity.services.storageService import StorageService
 from metadata.workflow.metadata import MetadataWorkflow
 from metadata.workflow.workflow_output_handler import print_status
 
 from ..containers import MinioContainerConfigs, get_minio_container
-from ..integration_base import OM_JWT, int_admin_ometa
 
 RESOURCES_DIR = Path(__file__).parent / "resources"
